+++ cjs/tests/index.test.ts
@@ -59,6 +59,29 @@ describe('CheqdSDK', () => {
 		);
 
 		it(
+			'it can dynamically instantiate Oracle module if available',
+			async () => {
+				const options = {
+					modules: [FeemarketModule as unknown as AbstractCheqdSDKModule],
+					rpcUrl: localnet.rpcUrl,
+					network: localnet.network,
+					wallet: await DirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic, { prefix: 'cheqd' }),
+				} satisfies ICheqdSDKOptions;
+				const cheqdSDK = await createCheqdSDK(options);
+
+				const sdkMethods = Object.keys(cheqdSDK.methods);
+				const testSigner = await CheqdSigningStargateClient.connectWithSigner(options.rpcUrl, options.wallet);
+				const testQuerier = (await CheqdQuerier.connectWithExtension(
+					options.rpcUrl,
+					setupOracleExtension
+				)) as CheqdQuerier & OracleExtension;
+
+				expect(sdkMethods).toContain('convertUSDtoCHEQ');
+			},
+			defaultAsyncTxTimeout
+		);
+
+		it(
 			'should use module methods',
 			async () => {
 				const rpcUrl = localnet.rpcUrl;